configs

+++ src/main/java/net/mehvahdjukaar/supplementaries/client/renderers/tiles/FlagBlockTileRenderer.java
@@ -85,7 +85,7 @@ public class FlagBlockTileRenderer implements BlockEntityRenderer<FlagBlockTile>
 
             BlockPos bp = tile.getBlockPos();
             //always from 0 to 1
-            float t = ((float) Math.floorMod((long) (bp.getX() * 7 + bp.getZ() * 13) + time, period) + partialTicks) / ((float) period);
+            float t = ((float) Math.floorMod(bp.getX() * 7L + bp.getZ() * 13L + time, period) + partialTicks) / ((float) period);
 
             if (ClientConfigs.cached.FLAG_BANNER) {
                 float ang = (float) ((wavyness + invdamping * w) * Mth.sin((float) ((((w / l) - t * 2 * (float) Math.PI)))));
